services/frame-reader/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cv-pipeline-frame-reader"
version = "0.1.0"
description = "GStreamer RTSP/file → NVDEC → Redis frame reader"
requires-python = ">=3.11"
dependencies = [
    "cv-pipeline-shared",
    "redis>=5.0",
]

[project.optional-dependencies]
gpu = [
    "torch>=2.2",
    "cupy-cuda12x>=13.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "black",
    "ruff",
    "mypy",
]

[tool.hatch.build.targets.wheel]
packages = ["frame_reader"]
